Loading screen for recording manager and limiting upload - refactored ApiClient into own file - storing database in hidden directory on ios - change database file name from "cords" to "cords.db"
